+++ NEWS.md
@@ -4,7 +4,7 @@
 * Fixes bug when including a uniform with no adjustment terms in the summarize_ds_models function (Issue #180)
 * dht2 can now deal with strata where there are no observations (Issue #194)
 * Users to pass a list of models to summarize_ds_models rather than passing them individually. (Issue #149)
-
+* Truncation distances greater than the largest cutpoint value for binned data are no longer permitted as these cause fitting issues. (Issue #175)
 
 # Distance 2.0.0